\name{getTopBlastHits}
\alias{getTopBlastHits}
\title{Return top blast hits}
\usage{
  getTopBlastHits(blastTableFile)
}
\arguments{
  \item{blastTableFile}{The name of a file of tabulated
  blast results.}
}
\value{
  A reduced blast table with one hit per query
}
\description{
  Auxillary function
}

