\docType{data}
\name{my.mlgt.Result}
\alias{my.mlgt.Result}
\title{An example \code{\link{mlgtResult}} object.}
\format{a \code{\link{mlgtResult}} object.}
\source{
  Package mlgt
}
\description{
  This is the result of running \code{\link{mlgt}} on the
  sample data given in the README.
}
\author{
  Dave T. Gerrard, 2012-04-01
}
\keyword{datasets}

